Refactor: split behavior nodes into a source file
